Create a header for TxId and TxHash
Summary: So that code that do not need CTransaction but still need to manipulate ids can import exactly what it needs.

Test Plan:
  make check

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1693
